Split the HtmlLicenseCreator file in two files
