+++ src/main/scala/nl/knaw/dans/easy/license/PlaceholderMapper.scala
@@ -37,7 +37,7 @@ import scala.collection.JavaConverters._
 import scala.language.{implicitConversions, postfixOps}
 import scala.util.Try
 
-class HtmlLicenseCreator(metadataTermsFile: File)(implicit parameters: Parameters) {
+class PlaceholderMapper(metadataTermsFile: File)(implicit parameters: Parameters) {
 
   val log = LoggerFactory.getLogger(getClass)
 
@@ -194,43 +194,5 @@ class HtmlLicenseCreator(metadataTermsFile: File)(implicit parameters: Parameter
     } yield map
 
     xs.foldLeft(new ju.ArrayList[ju.Map[String, String]])((list, map) => { list.add(map); list })
-  }
-}
-
-class VelocityTemplateResolver(propertiesFile: File)(implicit parameters: Parameters) {
-
-  val log = LoggerFactory.getLogger(getClass)
-
-  val properties = loadProperties(propertiesFile)
-    .doOnError(e => log.error(s"could not read the velocity properties in $propertiesFile", e))
-    .getOrElse(new ju.Properties())
-
-  val velocityResources = new File(properties.getProperty("file.resource.loader.path"))
-  val templateFileName = properties.getProperty("template.file.name")
-
-  val doc = new File(velocityResources, templateFileName)
-  assert(doc.exists(), s"file does not exist - $doc")
-
-  val engine = {
-    val engine = new VelocityEngine(properties)
-    engine.init()
-    engine
-  }
-
-  /**
-    * Create the template and write it to `out` after filling in the placeholders with `map`.
-    *
-    * @param out The `OutputStream` where the filled in template is written to
-    * @param map The mapping between placeholders and actual values
-    * @param encoding The encoding to be wused in writing to `out`
-    * @return `Success` if filling in the template succeeded, `Failure` otherwise
-    */
-  def createTemplate(out: OutputStream, map: PlaceholderMap, encoding: Charset = encoding) = {
-    new OutputStreamWriter(out).use(writer => {
-      val context = new VelocityContext
-      map.foreach { case (kw, o) => context.put(kw.keyword, o) }
-
-      engine.getTemplate(templateFileName, encoding.displayName()).merge(context, writer)
-    })
   }
 }
